Add checks for limit=0 to auto-generated endpoints.
Includes tests and doctest updates. I considered using the allow_limit_0
arg/attribute to also bypass flagging itemsPerPage=0 in get_paginated,
but I ran into complications when setting up the tests. Finally I
decided that it went against the entire point of a pagination method to
be able to bypass pagination and just added the flagging.