Doc updates using task management models/schemas.
One thing I want to note is that, in `docs/advanced/joins.md`, the
examples between the "Getting Joined Data Nested" section and the later
"One-to-One Relationships" appear to be identical? It feels
unnecessarily redundant to me, but I'm not sure what to do about it. I
just wanted to get it noted for the review and see if anyone else who
reads this has any ideas.